fix: move logout to profile, however won't clear cache for some reason
