+++ app/src/App.tsx
@@ -1,16 +1,23 @@
 /** @format */
 
-import React from 'react';
+import React, { useEffect, useState } from 'react';
+import { StatusBar } from 'react-native';
+// Screens
+import { SplashScreen } from '&screens';
+
+// Stack Navigators
+import { NavigationContainer } from '@react-navigation/native';
+import { AuthStackNavigator, OnboardingStackNavigator } from '&navigation';
 
 // Theme
 import { theme } from '&theme';
 import { ThemeProvider } from 'styled-components/native';
 
-// Navigation
-import { Root } from '&navigation';
-
 // Apollo
 import { ApolloClient, ApolloProvider, HttpLink, InMemoryCache } from '@apollo/client';
+
+// Firebase Authentication
+import auth, { FirebaseAuthTypes } from '@react-native-firebase/auth';
 
 // Context Store
 import AuthProvider from '&stores';
@@ -29,14 +36,36 @@ const client = new ApolloClient({
   }),
 });
 
-const App = () => (
-  <ThemeProvider theme={theme}>
-    <ApolloProvider client={client}>
-      <AuthProvider>
-        <Root />
-      </AuthProvider>
-    </ApolloProvider>
-  </ThemeProvider>
-);
+const App = () => {
+  const [loading, setLoading] = useState(true);
+  const [user, setUser] = useState<FirebaseAuthTypes.User | null>(null);
+
+  const onAuthStateChanged = (user: FirebaseAuthTypes.User | null) => {
+    setUser(user);
+    if (loading) setLoading(false);
+  };
+
+  useEffect(() => {
+    const subscriber = auth().onAuthStateChanged(onAuthStateChanged);
+    return subscriber;
+  }, []);
+
+  return (
+    <ThemeProvider theme={theme}>
+      <ApolloProvider client={client}>
+        <AuthProvider>
+          <>
+            <StatusBar barStyle="dark-content" />
+            {loading ? (
+              <SplashScreen />
+            ) : (
+              <NavigationContainer>{user ? <AuthStackNavigator /> : <OnboardingStackNavigator />}</NavigationContainer>
+            )}
+          </>
+        </AuthProvider>
+      </ApolloProvider>
+    </ThemeProvider>
+  );
+};
 
 export default App;
