+++ app/src/navigation/index.ts
@@ -1,5 +1,4 @@
 /** @format */
 
-export * from './Root';
 export * from './AuthStack';
 export * from './OnboardingStack';
