+++ app/src/screens/Auth/MapView/index.tsx
@@ -8,9 +8,6 @@ import { FullScreenWhite, HelloUser } from '&components';
 // Context
 import { AuthContext } from '&stores';
 
-// Firebase Authentication
-import auth from '@react-native-firebase/auth';
-
 // GraphQL
 import { useGetUserQuery } from '&graphql';
 
@@ -19,22 +16,5 @@ export const MapViewScreen = () => {
 
   const { data: authData } = useGetUserQuery({ variables: { sessionId: authContext?.token! } });
 
-  return (
-    <FullScreenWhite>
-      {authData && authContext && (
-        <HelloUser
-          userData={authData.User[0]}
-          logOut={() => {
-            void auth()
-              .signOut()
-              .then(
-                async () =>
-                  // @ts-ignore
-                  await authContext.removeToken()
-              );
-          }}
-        />
-      )}
-    </FullScreenWhite>
-  );
+  return <FullScreenWhite>{authData && authContext && <HelloUser userData={authData.User[0]} />}</FullScreenWhite>;
 };
